Separate out ESP/Teensy leadscrew io implementations
