+++ lib/leadscrew/leadscrew_io_esp.h
@@ -1,0 +1,27 @@
+
+#include <Wire.h>
+
+#include "leadscrew_io.h"
+#pragma once
+
+#ifdef ESP32
+
+class LeadscrewIOESP : public LeadscrewIO {
+  inline void writeStepPin(uint8_t val) {
+    if (val == 1)
+      REG_SET_BIT(GPIO_OUT_REG, ELS_LEADSCREW_STEP_BIT);
+    else
+      REG_CLR_BIT(GPIO_OUT_REG, ELS_LEADSCREW_STEP_BIT);
+    stepBitState = val;
+  }
+  inline uint8_t readStepPin() { return stepBitState; }
+
+  inline void writeDirPin(uint8_t val) {
+    if (val == 1)
+      REG_SET_BIT(GPIO_OUT_REG, ELS_LEADSCREW_DIR_BIT);
+    else
+      REG_CLR_BIT(GPIO_OUT_REG, ELS_LEADSCREW_DIR_BIT);
+  }
+  inline u_int8_t readDirPin() { return digitalRead(ELS_LEADSCREW_DIR); }
+};
+#endif
